# Fix performance issue of m-grouped contiguous GEMMs.